+++ src/components/Notes.css
@@ -1,9 +1,17 @@
 #notesDiv{
     display: flex;
-    justify-content: space-between;
+    flex-direction:column;
     background-color: #FEE4B7;
     height: 812px;
     padding: 2em;
+}
+
+#heatherSection{
+    border: 2px solid red;
+    display: flex;
+    flex-direction: row;
+    justify-content:space-between;
+    padding: 1em;
 }
 #logoutBtn{
     background-color: #053752; 
@@ -23,9 +31,39 @@
     font-family: 'Inter', sans-serif;
     font-weight: 400;
     margin-top: 10px;
-    margin-bottom: 1em;
+    margin-right: 1em;
     font-size: 1em;
     text-decoration: none;
+    min-width: 50px;
+}
+
+#newNotesSctn{
+    border: 2px solid black;
+    padding: 1em;
+    display: flex;
+    flex-direction: column;
+}
+#titleInput{
+    background-color: #fff7f7;
+    border: none;
+    width: 90%;
+}
+#noteInput{
+    background-color: #fff7f7;
+    border: none;
+    width: 90%;
+    height: 10em;
+}
+#saveBtn{
+    background-color: #053752; 
+    color:#FDF8DD;
+    font-family: 'Inter', sans-serif;
+    font-weight: 400;
+    font-size: 1em;
+    border-radius: 8px;
+    cursor: pointer;
+    padding: 0.6em 1.2em;
+    border: 1px solid transparent;
 }
 
 /* TABLET STYLE */
