+++ src/components/notes.jsx
@@ -1,13 +1,12 @@
 /* eslint-disable no-console */
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { getAuth, onAuthStateChanged } from 'firebase/auth';
-import { logOut } from '../firebase/Auth';
+import { onAuthStateChanged } from 'firebase/auth';
+import { auth, logOut } from '../firebase/firebase-init';
 import './Notes.css';
 
 function Notes() {
   const [name, setName] = useState('');
-  const auth = getAuth();
 
   onAuthStateChanged(auth, (user) => {
     const { displayName } = user;
@@ -27,15 +26,31 @@ function Notes() {
   };
 
   return (
+
     <div id="notesDiv">
-      <h2 id="welcome">
-        {' '}
-        Welcome to your notes
-        {' '}
-        {name}
-      </h2>
-      <button id="logoutBtn" type="button" onClick={signOut}> Sign out </button>
+      <section id="heatherSection">
+        <h2 id="welcome">
+          {' '}
+          Welcome to your notes
+          {' '}
+          {name}
+        </h2>
+        <button id="logoutBtn" type="button" onClick={signOut}> Sign out </button>
+      </section>
+
+      <section id="newNotesSctn">
+        <form>
+          <textarea id="titleInput" />
+          <textarea id="noteInput" />
+          <button id="saveBtn" type="submit"> Save </button>
+        </form>
+      </section>
+      <section>
+        <ul> </ul>
+      </section>
+
     </div>
+
   );
 }
